[recipe] mod: split the lib into targets per project
Split the library based on the project.
So now there is a main target which is LLVM.
And there are individual targets for each enabled projects. e.g. Clang, LLD etc

This is to fix recursive inclusion from different enabled projects that requires LLVM which triggers the recipe's LLVM target which then called that project's config and this repeated indefinitely.